AssetBundle: fix UnityWeb/UnityRaw header decomposition
UnityWeb and UnityRaw files are composed of two parts; a metadata for
the file itself and the bundle payload itself. That bundle itself also
has two parts, a header describing the assets that follow, and the
assets themselves.

Perform the entirety of parsing both metadata layers in AssetBundle,
leaving the Asset class to worry about only the assets themselves.